+++ tests/countingbloom_test.py
@@ -312,18 +312,18 @@ class TestCountingBloomFilter(unittest.TestCase):
         blm = CountingBloomFilter(est_elements=10, false_positive_rate=0.05)
         self.assertEqual(blm.elements_added, 0)
         res = blm.add('this is a test 0', large)
-        self.assertEqual(blm.elements_added, large - 1)
+        self.assertEqual(blm.elements_added, large)
         self.assertEqual(res, large - 1)
 
     def test_cbf_remove_from_large(self):
         ''' test adding a very large number of elements '''
-        large = 2 ** 32
+        large = 2**32
         blm = CountingBloomFilter(est_elements=10, false_positive_rate=0.05)
         self.assertEqual(blm.elements_added, 0)
         res = blm.add('this is a test 0', large)
-        self.assertEqual(blm.elements_added, large - 1)
+        self.assertEqual(blm.elements_added, large)
         self.assertEqual(res, large - 1)
 
         res = blm.remove('this is a test 0', large)
-        self.assertEqual(blm.elements_added, large - 1)
+        self.assertEqual(blm.elements_added, large)
         self.assertEqual(res, large - 1)